# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A REST API to interface with Sidero Omni.",
        "title": "Talos Omni Control API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/clustermachines": {
            "get": {
                "description": "Get a list of all cluster machines in Omni",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clustermachines"
                ],
                "summary": "List all cluster machines",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by cluster ID",
                        "name": "cluster",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.ClusterMachineResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/clustermachines/{id}/config-status": {
            "get": {
                "description": "Get configuration status information for a specific cluster machine",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clustermachines"
                ],
                "summary": "Get cluster machine config status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cluster Machine ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ClusterMachineConfigStatusResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/clustermachines/{id}/status": {
            "get": {
                "description": "Get detailed status information about a specific cluster machine",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clustermachines"
                ],
                "summary": "Get cluster machine status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cluster Machine ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ClusterMachineStatusResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/clusters": {
            "get": {
                "description": "Get a list of all clusters in Omni",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clusters"
                ],
                "summary": "List all clusters",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.ClusterResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/clusters/{id}/status": {
            "get": {
                "description": "Get health and phase information for a specific cluster",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "clusters"
                ],
                "summary": "Get cluster status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cluster ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.ClusterStatusResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/machines": {
            "get": {
                "description": "Get a list of all machines in Omni",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "machines"
                ],
                "summary": "List all machines",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.MachineResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/machines/{id}/status": {
            "get": {
                "description": "Get detailed status information about a specific machine. This endpoint is deprecated - status information is now included in the main machine endpoint.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "machines"
                ],
                "summary": "Get machine status (deprecated: use GET /machines/{id} instead)",
                "deprecated": true,
                "parameters": [
                    {
                        "type": "string",
                        "description": "Machine ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.MachineResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/schematics": {
            "get": {
                "description": "Get a list of all Talos image schematics in Omni",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "schematics"
                ],
                "summary": "List all schematics",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.SchematicResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.ClusterMachineConfigStatusResponse": {
            "type": "object",
            "properties": {
                "_links": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "cluster_machine_config_sha256": {
                    "type": "string"
                },
                "cluster_machine_config_version": {
                    "type": "string"
                },
                "cluster_machine_version": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_config_error": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                },
                "schematic_id": {
                    "type": "string"
                },
                "talos_version": {
                    "type": "string"
                }
            }
        },
        "handlers.ClusterMachineResponse": {
            "type": "object",
            "properties": {
                "_links": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "kubernetes_version": {
                    "type": "string"
                },
                "labels": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "machine_id": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                }
            }
        },
        "handlers.ClusterMachineStatusResponse": {
            "type": "object",
            "properties": {
                "_links": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "apid_available": {
                    "type": "boolean"
                },
                "config_apply_status": {
                    "type": "string"
                },
                "config_up_to_date": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "is_removed": {
                    "type": "boolean"
                },
                "last_config_error": {
                    "type": "string"
                },
                "management_address": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                },
                "provision_status": {
                    "$ref": "#/definitions/handlers.ProvisionStatus"
                },
                "ready": {
                    "type": "boolean"
                },
                "stage": {
                    "type": "string"
                }
            }
        },
...
        "handlers.ClusterResponse": {
            "type": "object",
            "properties": {
                "_links": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "features": {
                    "type": "object",
                    "properties": {
                        "disk_encryption": {
                            "type": "boolean"
                        },
                        "workload_proxy": {
                            "type": "boolean"
                        }
                    }
                },
                "id": {
                    "type": "string"
                },
                "kubernetes_version": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                },
                "talos_version": {
                    "type": "string"
                }
            }
        },
        "handlers.ClusterStatusResponse": {
            "type": "object",
            "properties": {
                "available": {
                    "type": "boolean"
                },
                "controlplane_ready": {
                    "type": "boolean"
                },
                "has_connected_control_planes": {
                    "type": "boolean"
                },
                "kubernetes_api_ready": {
                    "type": "boolean"
                },
                "machines": {
                    "type": "object",
                    "properties": {
                        "connected": {
                            "type": "integer"
                        },
                        "healthy": {
                            "type": "integer"
                        },
                        "requested": {
                            "type": "integer"
                        },
                        "total": {
                            "type": "integer"
                        }
                    }
                },
                "phase": {
                    "type": "string"
                },
                "ready": {
                    "type": "boolean"
                }
            }
        },
...
        "handlers.MachineResponse": {
            "type": "object",
            "properties": {
                "_links": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "arch": {
                    "type": "string"
                },
                "connected": {
                    "type": "boolean"
                },
                "hostname": {
                    "description": "Status fields (from MachineStatus resource)",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "labels": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "last_error": {
                    "type": "string"
                },
                "maintenance": {
                    "type": "boolean"
                },
                "management_address": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                },
                "platform": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "talos_version": {
                    "type": "string"
                },
                "use_grpc_tunnel": {
                    "type": "boolean"
                }
            }
        },
...
        "handlers.ProvisionStatus": {
            "type": "object",
            "properties": {
                "provider_id": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
...
        "handlers.SchematicResponse": {
            "type": "object",
            "properties": {
                "_links": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "namespace": {
                    "type": "string"
                }
            }
        },
...
    }
}